refactor: split action to separate file
